Set up in solidarity bot
Sets up a organization wide configuration of the in solidarity bot. This
bot will perform checks on changed code to look for usages of non-inclusive
terms, setting a failing status check for usages of whitelist, blacklist
and slave.

Note: A notice, rather than failure, is used for usages of master due to
our current usage of the term master for our primary branches.